+++ packages/python/plotly/plotly/validators/layout/hoverlabel/__init__.py
@@ -2,6 +2,7 @@ import sys
 
 if sys.version_info < (3, 7):
     from ._namelength import NamelengthValidator
+    from ._grouptitlefont import GrouptitlefontValidator
     from ._font import FontValidator
     from ._bordercolor import BordercolorValidator
     from ._bgcolor import BgcolorValidator
@@ -14,6 +15,7 @@ else:
         [],
         [
             "._namelength.NamelengthValidator",
+            "._grouptitlefont.GrouptitlefontValidator",
             "._font.FontValidator",
             "._bordercolor.BordercolorValidator",
             "._bgcolor.BgcolorValidator",